Closes #52.
