+++ phy/utils/array.py
@@ -6,6 +6,8 @@
 # Imports
 #------------------------------------------------------------------------------
 
+from math import floor
+
 import numpy as np
 
 from ..ext import six
@@ -14,6 +16,35 @@ from ..ext import six
 #------------------------------------------------------------------------------
 # Utility functions
 #------------------------------------------------------------------------------
+
+def _range_from_slice(myslice, start=None, stop=None, step=None, length=None):
+    """Convert a slice to an array of integers."""
+    assert isinstance(myslice, slice)
+    # Find 'step'.
+    step = step if step is not None else myslice.step
+    if step is None:
+        step = 1
+    # Find 'start'.
+    start = start if start is not None else myslice.start
+    if start is None:
+        start = 0
+    # Find 'stop' as a function of length if 'stop' is unspecified.
+    stop = stop if stop is not None else myslice.stop
+    if length is not None:
+        stop_inferred = floor(start + step * length)
+        if stop is not None and stop < stop_inferred:
+            raise ValueError("'stop' ({stop}) and ".format(stop=stop) +
+                             "'length' ({length}) ".format(length=length) +
+                             "are not compatible.")
+        stop = stop_inferred
+    if stop is None and length is None:
+        raise ValueError("'stop' and 'length' cannot be both unspecified.")
+    myrange = np.arange(start, stop, step)
+    # Check the length if it was specified.
+    if length is not None:
+        assert len(myrange) == length
+    return myrange
+
 
 def _unique(x):
     """Faster version of np.unique().
