+++ phy/utils/tests/test_array.py
@@ -12,13 +12,67 @@ from pytest import raises
 
 from ..array import (_unique, _normalize, _index_of, _as_array,
                      chunk_bounds, excerpts, data_chunk,
-                     PartialArray, _partial_shape)
+                     PartialArray, _partial_shape,
+                     _range_from_slice)
 from ...datasets.mock import artificial_spike_clusters
 
 
 #------------------------------------------------------------------------------
 # Test utility functions
 #------------------------------------------------------------------------------
+
+def test_range_from_slice():
+    """Test '_range_from_slice'."""
+
+    class _SliceTest(object):
+        """Utility class to make it more convenient to test slice objects."""
+        def __init__(self, **kwargs):
+            self._kwargs = kwargs
+
+        def __getitem__(self, item):
+            if isinstance(item, slice):
+                return _range_from_slice(item, **self._kwargs)
+
+    with raises(ValueError):
+        _SliceTest()[:]
+    with raises(ValueError):
+        _SliceTest()[1:]
+    ae(_SliceTest()[:5], [0, 1, 2, 3, 4])
+    ae(_SliceTest()[1:5], [1, 2, 3, 4])
+
+    with raises(ValueError):
+        _SliceTest()[::2]
+    with raises(ValueError):
+        _SliceTest()[1::2]
+    ae(_SliceTest()[1:5:2], [1, 3])
+
+    with raises(ValueError):
+        _SliceTest(start=0)[:]
+    with raises(ValueError):
+        _SliceTest(start=1)[:]
+    with raises(ValueError):
+        _SliceTest(step=2)[:]
+
+    ae(_SliceTest(stop=5)[:], [0, 1, 2, 3, 4])
+    ae(_SliceTest(start=1, stop=5)[:], [1, 2, 3, 4])
+    ae(_SliceTest(stop=5)[1:], [1, 2, 3, 4])
+    ae(_SliceTest(start=1)[:5], [1, 2, 3, 4])
+    ae(_SliceTest(start=1, step=2)[:5], [1, 3])
+    ae(_SliceTest(start=1)[:5:2], [1, 3])
+
+    ae(_SliceTest(length=5)[:], [0, 1, 2, 3, 4])
+    with raises(ValueError):
+        _SliceTest(length=5)[:3]
+    ae(_SliceTest(length=5)[:10], [0, 1, 2, 3, 4])
+    ae(_SliceTest(length=5)[:5], [0, 1, 2, 3, 4])
+    ae(_SliceTest(start=1, length=5)[:], [1, 2, 3, 4, 5])
+    ae(_SliceTest(start=1, length=5)[:6], [1, 2, 3, 4, 5])
+    with raises(ValueError):
+        _SliceTest(start=1, length=5)[:4]
+    ae(_SliceTest(start=1, step=2, stop=5)[:], [1, 3])
+    ae(_SliceTest(start=1, stop=5)[::2], [1, 3])
+    ae(_SliceTest(stop=5)[1::2], [1, 3])
+
 
 def test_unique():
     """Test _unique() function"""
